Removed a memory leak in CMedia.  Removed memory leak of OpenEXR.
Former-commit-id: ea735aa8fc3eed3511fd4cb20c4e5fe95b06b6cf [formerly ea735aa8fc3eed3511fd4cb20c4e5fe95b06b6cf [formerly ea735aa8fc3eed3511fd4cb20c4e5fe95b06b6cf [formerly ea735aa8fc3eed3511fd4cb20c4e5fe95b06b6cf [formerly ea735aa8fc3eed3511fd4cb20c4e5fe95b06b6cf [formerly ea735aa8fc3eed3511fd4cb20c4e5fe95b06b6cf [formerly d4ed2bf1c69ae0ee82ddc82b0bc3f1dcc02be520]]]]]]
Former-commit-id: 01a689b295f70b761e4d160e444451d81fe1d733 [formerly 7f1b41f22ad5395eb53b847cab622ccdb4871d7b] [formerly 68f02998aa5c5d599e2209ad115be2fdf5f831d6 [formerly 68f02998aa5c5d599e2209ad115be2fdf5f831d6 [formerly 68f02998aa5c5d599e2209ad115be2fdf5f831d6 [formerly 68f02998aa5c5d599e2209ad115be2fdf5f831d6 [formerly 68f02998aa5c5d599e2209ad115be2fdf5f831d6 [formerly 81d7454f1bcbd02dcebde08561b7dc713c7e3fd3]]]]]]
Former-commit-id: 68f02998aa5c5d599e2209ad115be2fdf5f831d6
Former-commit-id: c326a1c60d4a34bd7e1a5a53cec017714d62a6f0
Former-commit-id: ba50bb8b48be0eb5ea22b92c07a37bd6f77c1986
Former-commit-id: 5fd77c9e460c7aa65a159259b5908a3dd5d5fad8
Former-commit-id: 82040875aa211ed06e4e162ed61eaeffc98ea69e
Former-commit-id: 72ce8aeb7b32b993c80e218ac3d661973b39e00f

+++ mrViewer/src/gui/mrvVectorscope.cpp
@@ -66,7 +66,7 @@ namespace mrv
     int W2 = r.w() / 2;
     int H2 = r.h() / 2;
     int R2 = diameter_ / 2;
-    float angle = 32;
+    int angle = 32;
     for ( i = 0; i < 4; ++i, angle += 90 )
       {
 	fltk::push_matrix();
@@ -98,7 +98,7 @@ namespace mrv
       "G"
     };
 
-    angle = 15.0f;
+    angle = 15;
     for ( i = 0; i < 6; ++i, angle += 60 )
       {
 	fltk::push_matrix();
@@ -189,11 +189,11 @@ namespace mrv
             pic = img->right();
         if (!pic) return;
     }
-    if ( xmin >= (int)pic->width() ) xmin = pic->width()-1;
-    if ( ymin >= (int)pic->height() ) ymin = pic->height()-1;
+    if ( xmin >= pic->width() ) xmin = pic->width()-1;
+    if ( ymin >= pic->height() ) ymin = pic->height()-1;
 
-    if ( xmax >= (int)pic->width() ) xmax = pic->width()-1;
-    if ( ymax >= (int)pic->height() ) ymax = pic->height()-1;
+    if ( xmax >= pic->width() ) xmax = pic->width()-1;
+    if ( ymax >= pic->height() ) ymax = pic->height()-1;
 
 
     unsigned stepX = (xmax - xmin + 1) / w();
@@ -214,9 +214,9 @@ namespace mrv
     float one_gamma = 1.0f / gamma;
 
     CMedia::Pixel rp;
-    for ( unsigned y = (unsigned)ymin; (int) y <= ymax; y += stepY )
+    for ( unsigned y = ymin; y <= ymax; y += stepY )
       {
-          for ( unsigned x = (unsigned)xmin; (int) x <= xmax; x += stepX )
+	for ( unsigned x = xmin; x <= xmax; x += stepX )
 	  {
 	    CMedia::Pixel op = pic->pixel( x, y );
